requirements.txt
flask==2.3.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0
python-dotenv==1.0.0
pandas==2.1.3
werkzeug==2.3.7
gunicorn==21.2.0
